# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plog CMS API
  description: Plog CMS 系统 API 接口规范
  version: 1.0.0
  contact:
    name: Plog Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://admin.example.com/api/v1
    description: 生产环境
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: 获取文章列表
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: count
          in: query
          schema:
            type: integer
            default: 20
        - name: sort_id
          in: query
          schema:
            type: integer
        - name: keyword
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [date, views, comnum]
            default: date
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'

    post:
      tags: [Posts]
      summary: 创建文章
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCreateResponse'
  /posts/{id}:
    get:
      tags: [Posts]
      summary: 获取文章详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: password
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetailResponse'
        '404':
          description: 文章不存在
...
  /tags:
    get:
      tags: [Tags]
      summary: 获取标签列表
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: popular
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Tags]
      summary: 创建标签
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tagname]
              properties:
                tagname:
                  type: string
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /tags/{id}:
    get:
      tags: [Tags]
      summary: 获取标签详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Tags]
      summary: 更新标签
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tagname:
                  type: string
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Tags]
      summary: 删除标签
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: AUTH_COOKIE
    ApiKeyAuth:
      type: apiKey
      in: query
      name: api_key
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        cover:
          type: string
        url:
          type: string
        description:
          type: string
        date:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        author_avatar:
          type: string
        sort_id:
          type: integer
        sort_name:
          type: string
        views:
          type: integer
        comnum:
          type: integer
        like_count:
          type: integer
        top:
          type: string
        sortop:
          type: string
        tags:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              url:
                type: string
        need_pwd:
          type: string
        fields:
          type: object
        parent_id:
          type: integer
    PostListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            articles:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            page:
              type: integer
            total_pages:
              type: integer
            has_more:
              type: boolean
    PostDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            article:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string
                date:
                  type: string
                sort_id:
                  type: integer
                sort_name:
                  type: string
                type:
                  type: string
                author_id:
                  type: integer
                author_name:
                  type: string
                author_avatar:
                  type: string
                content:
                  type: string
                content_raw:
                  type: string
                excerpt:
                  type: string
                excerpt_raw:
                  type: string
                cover:
                  type: string
                views:
                  type: integer
                comnum:
                  type: integer
                like_count:
                  type: integer
                top:
                  type: string
                sortop:
                  type: string
                tags:
                  type: array
                fields:
                  type: object
                parent_id:
                  type: integer
    PostCreateRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        author_uid:
          type: integer
        post_date:
          type: string
        sort_id:
          type: integer
        tags:
          type: string
        cover:
          type: string
        draft:
          type: string
          enum: [y, n]
          default: n
        alias:
          type: string
        top:
          type: string
          enum: [y, n]
          default: n
        sortop:
          type: string
          enum: [y, n]
          default: n
        allow_remark:
          type: string
          enum: [y, n]
          default: y
        password:
          type: string
        link:
          type: string
        template:
          type: string
    PostCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            article_id:
              type: integer
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          type: object
          properties:
            request_id:
              type: string
              format: uuid
            timestamp:
              type: string
              format: date-time